requirements.txt
torch
numpy
torch
Pillow
requests
aiohttp
transformers
psutil
